Latest jas sources

+++ symja_android_library/matheclipse-external/src/main/java/edu/jas/arith/ModInt.java
@@ -51,7 +51,7 @@ public final class ModInt implements GcdRingElem<ModInt>, Modular {
    */
   public ModInt(ModIntRing m, long a) {
     ring = m;
-    int v = ((int) a) % ring.modul;
+    int v = (int) (a % (long) ring.modul);
     val = (v >= 0 ? v : v + ring.modul);
   }
 
@@ -72,7 +72,7 @@ public final class ModInt implements GcdRingElem<ModInt>, Modular {
    * @param a long.
    */
   public ModInt(ModIntRing m, Long a) {
-    this(m, a.intValue());
+    this(m, a.longValue());
   }
 
   /**
